fix: init react app
