--- rent-ripple/src/App.js
@@ -1,16 +1,0 @@
-import logo from './logo.png';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        {/* <h1>Rent Ripple</h1> */}
-        
-      </header>
-    </div>
-  );
-}
-
-export default App;
